requirements.txt
python-binance==1.0.16
pandas==1.5.3
numpy==1.24.3
python-telegram-bot==13.15
requests==2.28.2
python-dotenv==1.0.0
